# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_bitbiz_hias-core_domains_policy_schema.CreateUnderwritingRuleRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      is_blocking:
        type: boolean
      parameter_key:
        type: string
      parameter_value:
        type: string
      plan_id:
        type: string
      relationship:
        type: string
      risk_score_weight:
        type: integer
      rule_type:
        type: string
      severity:
        type: string
    required:
    - parameter_key
    - parameter_value
    - plan_id
    - rule_type
    type: object
...
  github_com_bitbiz_hias-core_domains_policy_schema.UpdateUnderwritingRuleRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      is_blocking:
        type: boolean
      parameter_key:
        type: string
      parameter_value:
        type: string
      relationship:
        type: string
      risk_score_weight:
        type: integer
      rule_type:
        type: string
      severity:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Health Insurance Administration System - Core API
  title: HIAS Core API
  version: "1.0"
paths:
...
  /api/v1/plans/{id}/underwriting-rules:
    get:
      consumes:
      - application/json
      description: List all underwriting rules for a plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List underwriting rules
      tags:
      - UnderwritingRules
    post:
      consumes:
      - application/json
      description: Create a new underwriting rule for a plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Rule details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.CreateUnderwritingRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an underwriting rule
      tags:
      - UnderwritingRules
...
  /api/v1/underwriting-rules/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an underwriting rule by ID
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an underwriting rule
      tags:
      - UnderwritingRules
    put:
      consumes:
      - application/json
      description: Update an existing underwriting rule by ID
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated rule details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.UpdateUnderwritingRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an underwriting rule
      tags:
      - UnderwritingRules
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"